requirements.txt
fastapi
uvicorn
gunicorn
requests
openai
python-dotenv
jinja2
python-multipart
pydantic
